refactor: form-decorator Moved to the `./.storybook` folder Switched to logging values directly, which is good for non-serializable things, such as File instances
