+++ .storybook/formDecorator.js
@@ -1,18 +1,15 @@
 import React from 'react'
 import { Button } from '@dhis2/ui-core'
-import { Form } from '../../src'
+import { Form } from '../src'
 
 const defaultFormProps = {
     onSubmit: values => {
         console.log(
-            [
-                '++++++++++++++++',
-                'Form was submitted with values:',
-                JSON.stringify(values, null, 4),
-                '----------------',
-            ].join('\n')
+            '++++++++++++++++\n',
+            'Form was submitted with values:\n',
+            values,
+            '\n----------------'
         )
-        console.log(values)
     },
     mutators: {},
 }
