+++ stories/CheckboxGroup.stories.js
@@ -8,7 +8,7 @@ import {
     arrayWithIdObjects,
     required,
 } from '../src'
-import { formDecorator } from './helpers/formDecorator'
+import { formDecorator } from '../.storybook/formDecorator'
 
 const options = [
     { label: 'Foo', value: 'foo' },
